Convert the rest of the major views into containers
